Making GetIncrementor const
Updating the GetIncrementor member function to be const to illustrate the point that a const member function was giving access to non-const members in this version of gcc. The 'this' pointer is decaying and loses its const-ness when captured by the lambda.

Also removing unnecessary semicolon. Not sure how that snuck its way in there.